Moved splits into split event, delegate to onStart
- Moved the splitting functionality out of update and into the split event. Removed doSplit in favour of multiple returns.

- Moved the isRelevant delegate creation to the onStart event, and stuffed it into vars.isRelevant. I didn't want to declare the delegate anew on every single frame. But it can't be declared in startup, because there, settings is still an ASLSettingsBuilder reference. onStart is the earliest time to capture it as an ASLSettingsReader reference.

If the delegate causes problems in any way in the future, the logic can just be inserted back to the two locations in situ.

Note also: the ASL script can no longer be changed on the fly while the timer is running, because of the delegate vars. But that shouldn't present a problem for normal use.