Simplify starter
The variable used for autostart is a field within CActionTitle. But it was being accessed in a very roundabout way, through another action which has a pointer to CActionTitle through the update linked list, then through CActionTitle's listener's pointer back to CActionTitle a second time. It now simply accesses CActionTitle directly.